Carry the cross-run stuck_count mean as f64 to stop fabricated stall regressions
The cross-run average fold rounded the aggregated stuck_count mean to
an integer. Across an A/B comparison pair the rounding error compounds
to a full unit, which does not clear the stuck_count metric's 1.0
absolute gate (the only rounded integer field that tight; the rest are
>= 5.0). So two cohorts averaging 1.4 and 1.6 stalls per run -- a true
difference of 0.2 -- rounded to 1 and 2 and were flagged as a
single-stall regression the underlying data never showed.

Widen the aggregated row's stuck_count from usize to f64 so it carries
the exact fractional mean and the comparison reads the true delta.
Other mean-folded fields (spread, migration_ratio, page_locality) are
already f64; this aligns stuck_count with them. The raw per-run monitor
count stays an integer, and the per-phase stuck_count carrier was
already f64 -- both unchanged.

Regression tests pin the fold's exact fractional mean and that the
comparison leaves a sub-integer difference unchanged while still
flagging a genuine whole-stall delta.

+++ src/stats/analyze.rs
@@ -22,7 +22,7 @@ const OUTLIER_METRICS: &[(&str, MetricAccessor)] = &[
     ("migration_ratio", |r| r.migration_ratio),
     ("imbalance", |r| r.imbalance_ratio),
     ("dsq_depth", |r| r.max_dsq_depth as f64),
-    ("stuck", |r| r.stuck_count as f64),
+    ("stuck", |r| r.stuck_count),
     ("fallback", |r| r.fallback_count as f64),
     ("keep_last", |r| r.keep_last_count as f64),
     // Distribution-kind roll-ups are ext_metrics-sourced (no typed field):
@@ -332,7 +332,7 @@ pub(crate) fn format_dimension_summary(rows: &[GauntletRow], group_col: &str) ->
             avg_dsq_depth: mean(group_rows.iter().map(|r| r.max_dsq_depth as f64)),
             total_stuck: group_rows
                 .iter()
-                .map(|r| r.stuck_count as f64)
+                .map(|r| r.stuck_count)
                 .filter(|x| x.is_finite())
                 .sum(),
             avg_fallback: mean(group_rows.iter().map(|r| r.fallback_count as f64)),
